Do not allow cross account sharing if bucket is encrypted with amazon managed key
