+++ backend/dataall/base/db/exceptions.py
@@ -141,13 +141,12 @@ class AWSInsufficientPrivileges(Exception):
     def __init__(self, action, message):
         self.action = action
         self.message = f"""
-                    An error occurred (AWSInsufficientPrivileges) when calling {self.action} operation. 
+                    An error occurred (AWSInsufficientPrivileges) when calling {self.action} operation.
                     Insufficient privileges for the pivot role: {message}
                 """
 
     def __str__(self):
         return f'{self.message}'
-
 
 
 class AWSResourceNotAvailable(Exception):
